Remove deprecated directive and upgrade to Angular 2.0.0 final
