+++ .gitignore
@@ -1,3 +1,3 @@
 bower_components
 node_modules
-directives/typings
+test/angular2/typings
