+++ package.json
@@ -25,12 +25,13 @@
     "install": "node_modules/.bin/bower install --allow-root"
   },
   "devDependencies": {
-    "@angular/common": "^2.0.0-rc.0",
-    "@angular/compiler": "^2.0.0-rc.0",
-    "@angular/core": "^2.0.0-rc.0",
-    "@angular/platform-browser": "^2.0.0-rc.0",
-    "@angular/platform-browser-dynamic": "^2.0.0-rc.0",
-    "@vaadin/angular2-polymer": "^1.0.0-alpha2",
+    "@angular/common": "^2.0.0",
+    "@angular/compiler": "^2.0.0",
+    "@angular/core": "^2.0.0",
+    "@angular/forms": "^2.0.0",
+    "@angular/platform-browser": "^2.0.0",
+    "@angular/platform-browser-dynamic": "^2.0.0",
+    "@vaadin/angular2-polymer": "^1.0.0-beta4",
     "bower": "latest",
     "chalk": "^1.1.1",
     "es6-shim": "^0.35.0",
@@ -42,11 +43,11 @@
     "gulp-typescript": "^2.13.0",
     "gulp-typings": "^1.3.4",
     "jshint": "^2.8.0",
-    "reflect-metadata": "0.1.2",
-    "rxjs": "5.0.0-beta.6",
+    "reflect-metadata": "0.1.3",
+    "rxjs": "5.0.0-beta.12",
     "systemjs": "0.19.27",
     "web-component-tester": "^4.1.0",
     "yargs": "^3.27.0",
-    "zone.js": "^0.6.12"
+    "zone.js": "^0.6.17"
   }
 }
