+++ test/angular2/app.module.js
@@ -1,0 +1,48 @@
+System.register(['@angular/core', '@angular/platform-browser', '@vaadin/angular2-polymer', './app.component'], function(exports_1, context_1) {
+    "use strict";
+    var __moduleName = context_1 && context_1.id;
+    var __decorate = (this && this.__decorate) || function (decorators, target, key, desc) {
+        var c = arguments.length, r = c < 3 ? target : desc === null ? desc = Object.getOwnPropertyDescriptor(target, key) : desc, d;
+        if (typeof Reflect === "object" && typeof Reflect.decorate === "function") r = Reflect.decorate(decorators, target, key, desc);
+        else for (var i = decorators.length - 1; i >= 0; i--) if (d = decorators[i]) r = (c < 3 ? d(r) : c > 3 ? d(target, key, r) : d(target, key)) || r;
+        return c > 3 && r && Object.defineProperty(target, key, r), r;
+    };
+    var __metadata = (this && this.__metadata) || function (k, v) {
+        if (typeof Reflect === "object" && typeof Reflect.metadata === "function") return Reflect.metadata(k, v);
+    };
+    var core_1, platform_browser_1, angular2_polymer_1, app_component_1;
+    var AppModule;
+    return {
+        setters:[
+            function (core_1_1) {
+                core_1 = core_1_1;
+            },
+            function (platform_browser_1_1) {
+                platform_browser_1 = platform_browser_1_1;
+            },
+            function (angular2_polymer_1_1) {
+                angular2_polymer_1 = angular2_polymer_1_1;
+            },
+            function (app_component_1_1) {
+                app_component_1 = app_component_1_1;
+            }],
+        execute: function() {
+            AppModule = (function () {
+                function AppModule() {
+                }
+                AppModule = __decorate([
+                    core_1.NgModule({
+                        imports: [platform_browser_1.BrowserModule],
+                        declarations: [app_component_1.TestApp, angular2_polymer_1.PolymerElement('vaadin-upload')],
+                        bootstrap: [app_component_1.TestApp],
+                        schemas: [core_1.CUSTOM_ELEMENTS_SCHEMA]
+                    }), 
+                    __metadata('design:paramtypes', [])
+                ], AppModule);
+                return AppModule;
+            }());
+            exports_1("AppModule", AppModule);
+        }
+    }
+});
+//# sourceMappingURL=app.module.js.map
