+++ test/angular2/typings.json
